app_a/requirements.txt
Flask==2.0.1
prometheus_client==0.11.0
werkzeug==2.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
